requirements.txt
Django==4.2.5
djangorestframework==3.14.0
model_bakery==1.18.1
